+++ e2e/helpers/pages/admin/members/members-page.ts
@@ -2,11 +2,7 @@ import {AdminPage} from '@/admin-pages';
 import {BasePage} from '@/helpers/pages';
 import {Download, JSHandle, Locator, Page} from '@playwright/test';
 import {readFileSync} from 'fs';
-
-export interface ExportedFile {
-    suggestedFilename: string;
-    content: string
-}
+import type {ExportedFile, MembersListSurface} from './members-list-page';
 
 class FilterSection extends BasePage {
     readonly actionsButton: Locator;
@@ -104,7 +100,7 @@ class SettingsSection extends BasePage {
     }
 }
 
-export class MembersPage extends AdminPage {
+export class MembersPage extends AdminPage implements MembersListSurface {
     readonly newMemberButton: Locator;
     public readonly loadMoreButton: Locator;
     public readonly membersListScrollRoot: Locator;
